feat: add UniswapV4DeployerCompetition (#117)
* feat: add UniswapV4DeployerCompetition

this commit adds a contract that creates a competition to generate the
vanity address for Uniswap V4. It does so using CREATE2 salts,
pre-calculating the address at which the contract will be deployed and
applying a score to the address based on its vanity. leading 0's are
weighted most heavily, followed by other 0's and 4's. The winner
receives an NFT, a bounty, and deployer privileges

* add comments

* change scoring to discuss

* interface plus tests

* inheritdoc tags

* change scoring algorithm

* feat: remove prizes

this commit removes the ETH payout, NFT mint, and exclusive deploy
rights

* feat: add natspec

* fix: snaps

* feat: improve test

* feat: cleanup scoring code

* fix: typo

* update constructor parameters for pool manager

* correct snapshots

* Include address in salt

* remove console logs

* correct test name

* More constructor parameters

* fix: remove override keywords

* fix: remove unused code

* fix: use default create2 function signature

* feat: remove unused import

* fix: remove unused v4Owner variable

* fix: add rules to natspec

* fix(deployComp): remove bestaddress storage

This commit removes the storage variable for bestAddres, calculating it
on-the-fly instead. It also just takes the initial `salt = 0` value as
a default score-to-beat rather than a sentinel for any salt is allowed
as a simplification

* fix: minor nits

- comparison strictness match comments
- unchecked vanity math

* snapshot

---------

Co-authored-by: dianakocsis <diana.kocsis@uniswap.org>
Co-authored-by: hensha256 <henshawalice@gmail.com>